Character Audio Setup
Scriptable Objects, SOCharacterAudio, and SOCharacterAudioPool, contain audio clips, character names, and info/logic about when to trigger a specific Sfxbase audio clip.

SOCharacterAudio and DialogueAudio should be modified to customize the triggering mechanism for sfx. 

TODO: The actual playing of the sound effect still needs to be hooked into the AudioManager.