Final Test Solution
I finally figured out when exporting the game exporter file and going into the fbx in unity and duplicating (CTRL+D) the motion file to extract it from the fbx actually messes up the naming system. By double clicking the names in the animation menu you are able to rename the animation properties to the correct name and Unity will automatically reconnect them to the asset.