Adjustable Brightness/Contrast and ScreenMode
Brightness/Contrast will only work on scenes with the GraphicsController attached to an object with a global Volume. Each level has been updated to include except main menu. We can setup of a volume for the menu or implement the ability to adjust the default profile.

Hooked Graphics Controller and FontManager into DataManager.SetVisualSettings