Lighting Optimization Phase 1
Reduced range of main lights and adjusted their position for now. Result is a little moodier lighting which looks just fine, but it can be adjusted if more visibility is needed. Reduced number of lights on lamps, and removed swarm of lights in Morgan's room with hopes of getting in emissive lights instead.