Gave narrator a voice. If you like the sound of it, maybe let him speak
Tried to implement dialogue triggered audio, but I was creating a ton of bugs and couldn't make it work. Made voices a BIT more recognizable. Spent a WHILE trying to figure out some audio bugs but couldn't so I guess they are going in lol. Listen to the letter in spring, if you like the sound of that, could you make the narrator sound like that?