Fixed Val's cracking voice.
Because of the nature of the randomizer, I can never TRULY get rid of it, so I'm calling the other voices acceptable, but Val was a menace to society and she had to be stopped. We had a stern talking and she's agreed to cooperate for now. Keep an eye on her.